+++ netket/hilbert/__init__.py
@@ -12,13 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from . import constraint
+from . import index
+
 from .abstract_hilbert import AbstractHilbert
 from .discrete_hilbert import DiscreteHilbert
 from .homogeneous import HomogeneousHilbert
 
 from .continuous_hilbert import ContinuousHilbert
 
-from .custom_hilbert import CustomHilbert
 from .doubled_hilbert import DoubledHilbert
 from .spin import Spin
 from .fock import Fock
@@ -28,6 +30,25 @@ from .particle import Particle
 from .tensor_hilbert import TensorHilbert
 from . import tensor_hilbert_discrete
 
+from . import random
+
+
+# Deprecated bindings
+from .custom_hilbert import CustomHilbert as _deprecated_CustomHilbert
+
+_deprecations = {
+    # September 2024, NetKet 3.14
+    "CustomHilbert": (
+        "netket.hilbert.CustomHilbert is deprecated: use custom constraints with "
+        "existing hilbert spaces instead, or define your own hilbert space class.",
+        _deprecated_CustomHilbert,
+    ),
+}
+
+from netket.utils.deprecation import deprecation_getattr as _deprecation_getattr
 from netket.utils import _hide_submodules
 
+__getattr__ = _deprecation_getattr(__name__, _deprecations)
 _hide_submodules(__name__)
+
+del _deprecation_getattr