+++ netket/hilbert/homogeneous.py
@@ -16,19 +16,62 @@ from collections.abc import Callable
 
 import numpy as np
 
+import jax
 import jax.numpy as jnp
 
 from equinox import error_if
 
-from netket.utils import StaticRange
+from netket.errors import InvalidConstraintInterface, UnhashableConstraintError
+from netket.utils import StaticRange, warn_deprecation
 from netket.utils.types import Array
 
 from .discrete_hilbert import DiscreteHilbert
+from .constraint import DiscreteHilbertConstraint
 from .index import (
     HilbertIndex,
     UniformTensorProductHilbertIndex,
     optimalConstrainedHilbertindex,
 )
+
+
+def check_and_deprecate_constraint(
+    constraint,
+    constraint_fn,
+):
+    __tracebackhide__ = True
+
+    if constraint is not None and constraint_fn is not None:
+        raise ValueError(
+            "Cannot specify at the same time `constraint` and `constraint_fn`, which"
+            "has been deprecated. Please remove the latter."
+        )
+    elif constraint_fn is not None and constraint is None:
+        constraint = constraint_fn
+        warn_deprecation(
+            "The keyword argument `constraint_fn` was renamed to `constraint` and "
+            "deprecated. Moreover, you can no longer pass a function, but must specify a class "
+            "according to the documentation."
+        )
+
+    if constraint is None:
+        return constraint
+
+    # now
+    if not isinstance(constraint, DiscreteHilbertConstraint):
+        raise InvalidConstraintInterface()
+
+    # Check that the constraint is well behaved and is jax-friendly hashable
+    try:
+        leaves, struct = jax.tree_util.tree_flatten(constraint)
+        if not len(leaves) == 0:
+            raise TypeError
+        hash(constraint)
+        constraint == constraint
+
+    except Exception as err:
+        raise UnhashableConstraintError(constraint) from err
+
+    return constraint
 
 
 class HomogeneousHilbert(DiscreteHilbert):
@@ -58,6 +101,8 @@ class HomogeneousHilbert(DiscreteHilbert):
         self,
         local_states: StaticRange | None,
         N: int = 1,
+        *,
+        constraint: DiscreteHilbertConstraint | None = None,
         constraint_fn: Callable | None = None,
     ):
         r"""
@@ -72,9 +117,11 @@ class HomogeneousHilbert(DiscreteHilbert):
                 numbers used to encode the local degree of freedom of this Hilbert
                 Space.
             N: Number of modes in this hilbert space (default 1).
-            constraint_fn: A function specifying constraints on the quantum numbers.
+            constraint: A callable class specifying constraints on the quantum numbers.
                 Given a batch of quantum numbers it should return a vector of bools
-                specifying whether those states are valid or not.
+                specifying whether those states are valid or not. It must be an
+                hashable subclass of :class:`netket.hilbert.constraint.DiscreteHilbertConstraint`.
+
         """
         assert isinstance(N, int)
 
@@ -85,7 +132,7 @@ class HomogeneousHilbert(DiscreteHilbert):
         self._local_size = len(local_states)
         self._is_finite = self._local_size < np.iinfo(np.intp).max
 
-        self._constraint_fn = constraint_fn
+        self._constraint = check_and_deprecate_constraint(constraint, constraint_fn)
 
         self._hilbert_index_ = None
 
@@ -115,6 +162,10 @@ class HomogeneousHilbert(DiscreteHilbert):
 
     def states_at_index(self, i: int):
         return self.local_states
+
+    @property
+    def constraint(self):
+        return self._constraint
 
     @property
     def n_states(self) -> int:
@@ -190,7 +241,7 @@ class HomogeneousHilbert(DiscreteHilbert):
         Typically, objects defined in the constrained space cannot be
         converted to QuTiP or other formats.
         """
-        return self._constraint_fn is not None
+        return self.constraint is not None
 
     def _numbers_to_states(self, numbers: np.ndarray) -> np.ndarray:
         return self._hilbert_index.numbers_to_states(numbers)
@@ -215,7 +266,7 @@ class HomogeneousHilbert(DiscreteHilbert):
         if self.constrained:
             states = error_if(
                 states,
-                ~self._constraint_fn(states).all(),
+                ~self.constraint(states).all(),
                 "States do not fulfill constraint.",
             )
 
@@ -248,7 +299,7 @@ class HomogeneousHilbert(DiscreteHilbert):
                 index = UniformTensorProductHilbertIndex(self._local_states, self.size)
             else:
                 index = optimalConstrainedHilbertindex(
-                    self._local_states, self.size, self._constraint_fn
+                    self._local_states, self.size, self.constraint
                 )
             self._hilbert_index_ = index
         return self._hilbert_index_
@@ -273,5 +324,5 @@ class HomogeneousHilbert(DiscreteHilbert):
             self.local_size,
             self._local_states,
             self.constrained,
-            self._constraint_fn,
+            self._constraint,
         )