+++ CHANGELOG.md
@@ -5,9 +5,14 @@
 
 ## NetKet 3.14 (⚙️ In development)
 
+### New features
+* Hilbert spaces such as {class}`nk.hilbert.Spin` and {class}`nk.hilbert.Fock`, as well as their base class {class}`nk.hilbert.HomogeneousHilbert`, now support arbitrary custom constraints [#1908](https://github.com/netket/netket/pull/1908).
+* The constraint interface has been stabilised, documented, and made compatible with several utilities. It is now possible to generate random states from arbitrary constrained hilbert spaces automatically, and it is possible to index into those spaces efficiently. Look at the hilbert space documentation for more information [#1908](https://github.com/netket/netket/pull/1908).
+
 ### Changes
 * Jax operators now use the same `chunk_size` as specified by the user when computing the forward pass. Prior to this change, Jax operators would be chunking the sample axis, but if an operator had a lot of connected elements this would end up increasing the effective sample size [#1875](https://github.com/netket/netket/pull/1875).
 * Metropolis Hamiltonian sampler for numba operators has been greatly simplified in order to remove the dependency on numba4jax. The new implementation will generally be slower than before, so we encourage you to use Jax Operators if possible. In the future, if people ask for it, we may reintroduce this implementation as a separate package [#1747](https://github.com/netket/netket/pull/1747).
+* Previously-internal hilbert space constraint sub-module located at `netket.hilbert.index.constraints` has been moved to `netket.hilbert.constraint` [#1908](https://github.com/netket/netket/pull/1908).
 
 ### Improvements
 * Specialised lattice constructors like {func}`nk.graph.Grid` now accept a `point_group` argument, overriding the default (usually maximal) point groups [#1879](https://github.com/netket/netket/pull/1879).